Refactored widget ui code into 2 separate files, one for the 2x2 and other to the 4x2
