+++ packages/uni_app/android/app/src/main/kotlin/pt/up/fe/ni/uni/receivers/ScheduleWidgetWideReceiver.kt
@@ -1,8 +1,8 @@
 package pt.up.fe.ni.uni.receivers
 
 import es.antonborri.home_widget.HomeWidgetGlanceWidgetReceiver
-import pt.up.fe.ni.uni.widgets.ScheduleWidget
+import pt.up.fe.ni.uni.widgets.ScheduleWideWidget
 
-class ScheduleWidgetWideReceiver : HomeWidgetGlanceWidgetReceiver<ScheduleWidget>() {
-    override val glanceAppWidget = ScheduleWidget()
+class ScheduleWidgetWideReceiver : HomeWidgetGlanceWidgetReceiver<ScheduleWideWidget>() {
+    override val glanceAppWidget = ScheduleWideWidget()
 }
