+++ packages/uni_app/android/app/src/main/kotlin/pt/up/fe/ni/uni/widgets/ScheduleWidget.kt
@@ -40,7 +40,7 @@ import com.google.gson.reflect.TypeToken
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
-private object WidgetTheme {
+internal object WidgetTheme {
     val Background = ColorProvider(day = Color(0xFFFFF5F3), night = Color(0xFF2F0A0C))
     val PrimaryText = ColorProvider(day = Color(0xFF000000), night = Color(0xFFFFFFFF))
     val SecondaryText = ColorProvider(day = Color(0xFF888888), night = Color(0xFFE5C8C7))
@@ -55,19 +55,9 @@ private object WidgetTheme {
 class ScheduleWidget : GlanceAppWidget() {
     override val stateDefinition: GlanceStateDefinition<*> = HomeWidgetGlanceStateDefinition()
 
-    override val sizeMode: SizeMode = SizeMode.Responsive(
-        setOf(
-            DpSize(110.dp, 110.dp), // Small (2x2)
-            DpSize(250.dp, 110.dp)  // Wide (4x2)
-        )
-    )
+    override val sizeMode: SizeMode = SizeMode.Single
 
-    override val previewSizeMode: PreviewSizeMode = SizeMode.Responsive(
-        setOf(
-            DpSize(110.dp, 110.dp), // Small (2x2)
-            DpSize(250.dp, 110.dp)  // Wide (4x2)
-        )
-    )
+    override val previewSizeMode: PreviewSizeMode = SizeMode.Single
 
     override suspend fun providePreview(context: Context, widgetCategory: Int) {
         val sampleLectures = listOf(
@@ -90,18 +80,13 @@ class ScheduleWidget : GlanceAppWidget() {
         )
 
         provideContent {
-            val size = androidx.glance.LocalSize.current
-            if (size.width >= 200.dp) {
-                WideWidgetUI(sampleLectures)
-            } else {
-                val nextLecture = sampleLectures.firstOrNull()
-                val acronym = nextLecture?.let { it["acronym"] as? String }
-                val subject = nextLecture?.let { it["subject"] as? String }
-                val typeClass = nextLecture?.let { it["typeClass"] as? String }
-                val room = nextLecture?.let { it["room"] as? String }
-                val startTimeStr = nextLecture?.let { it["startTime"] as? String }
-                WidgetUI(acronym, subject, typeClass, room, startTimeStr)
-            }
+            val nextLecture = sampleLectures.firstOrNull()
+            val acronym = nextLecture?.let { it["acronym"] as? String }
+            val subject = nextLecture?.let { it["subject"] as? String }
+            val typeClass = nextLecture?.let { it["typeClass"] as? String }
+            val room = nextLecture?.let { it["room"] as? String }
+            val startTimeStr = nextLecture?.let { it["startTime"] as? String }
+            WidgetUI(acronym, subject, typeClass, room, startTimeStr)
         }
     }
 
@@ -130,18 +115,13 @@ class ScheduleWidget : GlanceAppWidget() {
                 endTime?.isAfter(now) ?: false
             }
 
-            val size = androidx.glance.LocalSize.current
-            if (size.width >= 200.dp) {
-                WideWidgetUI(lectures)
-            } else {
-                val nextLecture = lectures.firstOrNull()
-                val acronym = nextLecture?.let { it["acronym"] as? String }
-                val subject = nextLecture?.let { it["subject"] as? String }
-                val typeClass = nextLecture?.let { it["typeClass"] as? String }
-                val room = nextLecture?.let { it["room"] as? String }
-                val startTimeStr = nextLecture?.let { it["startTime"] as? String }
-                WidgetUI(acronym, subject, typeClass, room, startTimeStr)
-            }
+            val nextLecture = lectures.firstOrNull()
+            val acronym = nextLecture?.let { it["acronym"] as? String }
+            val subject = nextLecture?.let { it["subject"] as? String }
+            val typeClass = nextLecture?.let { it["typeClass"] as? String }
+            val room = nextLecture?.let { it["room"] as? String }
+            val startTimeStr = nextLecture?.let { it["startTime"] as? String }
+            WidgetUI(acronym, subject, typeClass, room, startTimeStr)
         }
     }
 
@@ -263,204 +243,7 @@ class ScheduleWidget : GlanceAppWidget() {
         }
     }
 
-    @SuppressLint("RestrictedApi")
-    @Composable
-    fun WideWidgetUI(lectures: List<Map<String, Any>>) {
-        val context = LocalContext.current
-
-        Box(
-            modifier = GlanceModifier
-                .fillMaxSize()
-                .appWidgetBackground()
-                .background(WidgetTheme.Background)
-                .cornerRadius(R.dimen.widget_radius)
-                .padding(horizontal = 12.dp, vertical = 12.dp)
-                .clickable(actionStartActivity(Intent(context, MainActivity::class.java))),
-            contentAlignment = Alignment.TopStart
-        ) {
-            if (lectures.isEmpty()) {
-                Box(modifier = GlanceModifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-                    Text(text = "No upcoming classes", style = TextStyle(color = WidgetTheme.SecondaryText))
-                }
-            } else {
-                Column(modifier = GlanceModifier.fillMaxSize()) {
-                    lectures.take(2).forEachIndexed { index, lecture ->
-                        Row(
-                            modifier = GlanceModifier.fillMaxWidth().defaultWeight(),
-                            verticalAlignment = Alignment.Top
-                        ) {
-                            // Timeline Item for this lecture
-                            val startTimeStr = lecture["startTime"] as? String
-                            val displayTime = formatTime(startTimeStr)
-
-                            // Time Text Column
-                            Column(
-                                modifier = GlanceModifier.width(50.dp).fillMaxHeight(),
-                                horizontalAlignment = Alignment.End
-                            ) {
-                                Box(
-                                    modifier = GlanceModifier.fillMaxWidth().height(24.dp),
-                                    contentAlignment = Alignment.CenterEnd
-                                ) {
-                                    Text(
-                                        text = displayTime,
-                                        style = TextStyle(
-                                            fontSize = 16.sp,
-                                            fontWeight = FontWeight.Bold,
-                                            color = WidgetTheme.Accent
-                                        )
-                                    )
-                                }
-                            }
-
-                            Spacer(modifier = GlanceModifier.width(10.dp))
-
-                            // Dot + Line Column (Guarantees connection)
-                            Box(
-                                modifier = GlanceModifier.fillMaxHeight().width(22.dp),
-                                contentAlignment = Alignment.TopCenter
-                            ) {
-                                // Line (behind)
-                                Column(
-                                    modifier = GlanceModifier.fillMaxSize(),
-                                    horizontalAlignment = Alignment.CenterHorizontally
-                                ) {
-                                    Spacer(modifier = GlanceModifier.height(12.dp)) // To the center of the 24dp dot box
-                                    Box(
-                                        modifier = GlanceModifier
-                                            .width(3.dp)
-                                            .defaultWeight()
-                                            .background(WidgetTheme.Accent)
-                                    ) {}
-                                }
-
-                                // Dot (front)
-                                Box(
-                                    modifier = GlanceModifier.fillMaxWidth().height(24.dp),
-                                    contentAlignment = Alignment.Center
-                                ) {
-                                    Box(
-                                        modifier = GlanceModifier
-                                            .size(17.dp)
-                                            .background(WidgetTheme.Accent)
-                                            .cornerRadius(8.dp),
-                                        contentAlignment = Alignment.Center
-                                    ) {
-                                        Box(
-                                            modifier = GlanceModifier
-                                                .size(10.dp)
-                                                .background(WidgetTheme.Background)
-                                                .cornerRadius(5.dp)
-                                        ) {}
-                                    }
-                                }
-                            }
-
-                            Spacer(modifier = GlanceModifier.width(8.dp))
-
-                            // Lecture Card
-                            Box(modifier = GlanceModifier.defaultWeight()) {
-                                LectureCard(lecture)
-                            }
-                        }
-
-                        if (index == 0 && lectures.size > 1) {
-                            Spacer(modifier = GlanceModifier.height(12.dp))
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    @Composable
-    private fun LectureCard(
-        lecture: Map<String, Any>
-    ) {
-        val acronym = lecture["acronym"] as? String ?: ""
-        val subject = lecture["subject"] as? String ?: ""
-        val typeClass = lecture["typeClass"] as? String ?: ""
-        val room = lecture["room"] as? String ?: ""
-        val badgeColor = WidgetTheme.getBadgeColor(typeClass)
-
-        Box(
-            modifier = GlanceModifier
-                .fillMaxWidth()
-                .background(WidgetTheme.CardBackground)
-                .cornerRadius(24.dp)
-                .padding(horizontal = 16.dp, vertical = 16.dp)
-        ) {
-            Column(modifier = GlanceModifier.fillMaxWidth()) {
-                Row(
-                    modifier = GlanceModifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        text = acronym,
-                        style = TextStyle(
-                            fontSize = 24.sp,
-                            fontWeight = FontWeight.Bold,
-                            color = WidgetTheme.PrimaryText
-                        )
-                    )
-                    Spacer(modifier = GlanceModifier.width(10.dp))
-                    Box(
-                        modifier = GlanceModifier
-                            .background(badgeColor)
-                            .cornerRadius(100.dp)
-                            .padding(horizontal = 10.dp, vertical = 3.dp),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Text(
-                            text = typeClass,
-                            style = TextStyle(
-                                fontSize = 12.sp,
-                                fontWeight = FontWeight.Medium,
-                                color = ColorProvider(day = Color.White, night = Color.White)
-                            )
-                        )
-                    }
-                    Spacer(modifier = GlanceModifier.defaultWeight())
-                    // Location
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Box(
-                            modifier = GlanceModifier
-                                .size(24.dp)
-                                .background(WidgetTheme.Accent)
-                                .cornerRadius(12.dp),
-                            contentAlignment = Alignment.Center
-                        ) {
-                            Image(
-                                provider = ImageProvider(R.drawable.ic_pin),
-                                contentDescription = "Location",
-                                modifier = GlanceModifier.size(14.dp)
-                            )
-                        }
-                        Spacer(modifier = GlanceModifier.width(6.dp))
-                        Text(
-                            text = room,
-                            style = TextStyle(
-                                fontSize = 18.sp,
-                                fontWeight = FontWeight.Bold,
-                                color = WidgetTheme.PrimaryText
-                            )
-                        )
-                    }
-                }
-                Spacer(modifier = GlanceModifier.height(4.dp))
-                Text(
-                    text = subject,
-                    style = TextStyle(
-                        fontSize = 14.sp,
-                        color = WidgetTheme.SecondaryText
-                    ),
-                    maxLines = 1
-                )
-            }
-        }
-    }
-
-    private fun parseDateTime(dateTimeStr: String?): LocalDateTime? {
+    internal fun parseDateTime(dateTimeStr: String?): LocalDateTime? {
         if (dateTimeStr == null) return null
         return try {
             // Flutter's DateTime.toString() returns "YYYY-MM-DD HH:MM:SS.mmm"
@@ -472,7 +255,7 @@ class ScheduleWidget : GlanceAppWidget() {
         }
     }
 
-    private fun formatTime(startTimeStr: String?): String {
+    internal fun formatTime(startTimeStr: String?): String {
         return try {
             if (startTimeStr != null) {
                 val timePart = startTimeStr.split(" ")[1]
